Create test helpers file and update another test
