+++ packages/deployer/test/errors.js
@@ -6,7 +6,9 @@ const EventEmitter = require("events");
 const Deployer = require("../index");
 const utils = require("./helpers/utils");
 
-describe("Error cases", function () {
+const { mockEventsSystem } = require("./helpers");
+
+describe("Error cases", function() {
   let owner;
   let accounts;
   let options;
@@ -30,10 +32,6 @@ describe("Error cases", function () {
     },
     logging: { quiet: true }
   });
-
-  const mockMigration = {
-    emitter: new EventEmitter()
-  };
 
   const web3 = new Web3(provider);
 
@@ -63,6 +61,8 @@ describe("Error cases", function () {
     Abstract = utils.getContract("Abstract", provider, networkId, owner);
     Loops = utils.getContract("Loops", provider, networkId, owner);
 
+    mockEventsSystem.clearEmittedEvents();
+
     options = {
       contracts: [
         Example,
@@ -79,7 +79,8 @@ describe("Error cases", function () {
       },
       network: "test",
       network_id: networkId,
-      provider
+      provider,
+      events: mockEventsSystem
     };
     deployer = new Deployer({ options });
   });
@@ -100,9 +101,9 @@ describe("Error cases", function () {
       await deployer.start();
       assert.fail();
     } catch (err) {
-      assert(err.message.includes("Deployment Failed"));
-      assert(err.message.includes("IsLibrary"));
-      assert(err.message.includes("has no address"));
+      assert(mockEventsSystem.emittedEvents["migrate:deployment:error"].length === 1);
+      assert(mockEventsSystem.emittedEvents["migrate:deployment:error"][0].data.type === "noLibAddress");
+      assert(mockEventsSystem.emittedEvents["migrate:deployment:error"][0].data.contract.contractName === "IsLibrary");
     }
   });
